Add initial v1alpha2 conversion
Much of this is autogenerated by conversion-gen, with the remaining
manual conversions at least stubbed out so it compiles. There are
many TODOs remaining, mainly around the VirtualMachine Spec/Status
since that has changed the most in v1a2. We'll likely have to resort
to the CAPI trick of sorting the prior version deser to JSON in the
object's annotations to provide full v1a1<->v1a2. Note that from
some of our types, based on how they are used and who - us - creates
and updates them we can likely get by with some amount of lossy
conversions.

Add generate-go-conversions make target to generate the
conversions. Not hooked up to the generate target yet since it
takes a long time to run, and the Github runner env is different
enough in that the determined package names in some comments are
different, and the changes to our CI runner are kind of large &
annoying.

Import conversion fuzz testing from CAPI. Include first round of
fuzz tests with notes as to what is needed for lossless
conversions. Tests will be hooked in to the build later.