Support ControllerName for VM Class
This patch adds support for specifying the field `spec.controllerName`
in a `VirtualMachineClass` resource. If this field is set to
`vmoperator.vmware.com/vsphere`, then the VM Operator controller for
`VirtualMachine` reconciles all VM resources. If the field is empty,
then the environment variable `DEFAULT_VM_CLASS_CONTROLLER_NAME` is
used to determine the default controller name. If this environment
variable is empty, it defaults to `vmoperator.vmware.com/vsphere`.
Otherwise the controller responsible for reconciling VM resources is
determined by the field `spec.controllerName` in the VM Class
referenced by a VM.