Move remaining webhooks under v1alpha1 sub packages
Note that in v1a2 the WebConsoleRequest was renamed to
VirtualMachineWebConsoleRequest to have the same prefix as our other
types. Rename the top-level directory now to keep these types together.
We'll need to register both webhooks as long as we want to support the
old CR with the old name name.

For some types we'll need to keep the v1a1 mutation webhooks registered
even after we switch to v1a2 as the storage version.