Use Symlink/Docker for generate-go-conversions fallbacks
This patch updates the generate-go-conversions target to support
either Symlink (default) or Docker as a fallback for running the
conversion-gen tool when the local sources are not cloned in the
path expected by said tool.

The env var CONVERSION_GEN_FALLBACK_MODE may be used to set the
fallback mode. This defaults to symlink, but can also be set to
docker.