Import v1a2 pkg changes
This contains the non provider pkg/ changes for v1a2, which is most
of the changes required to get the v1a2 controllers working.

With the change to metav1.Condition types, the Reason field is now
required which previously we wouldn't provide when setting the
condition to true. To keep the later changes smaller we reuse the
type field value but will need to refactor our conditions API later
to add a reason argument and meaningful reason messages where we
want to.