Add v1a2 controllers
This imports the v1a2 controllers and other requisite changes. Like as
was done with the v1a2 webhooks in 9869e81, the integrations tests
disabled until we can sort out the mess of dealing with multiple
manifest generations on a single branch. Generally, when possible
this is mostly a mechanical change, and the diff between v1a1 and
v1a2 controllers is relatively small.

This is primarily diff reduction with the v1a2 branch, and allows
changes to be made to both versions together since we're still some
time from flipping the switch. This is unfortunately a painful and
disruptive change.