Bump manager container memory limit
The is the quickest change to alleviate the OOM we're seeing in
large scale test envs.

Longer term work of setting GOMEMLIMIT, using more metadata only
watches, disabling the ctrl-runtime Client for more types, and
reworking some types to not store large dumps of data in the
annotations so MD-only watches are more effective is needed.