Better handle when CnsNodeVmAttachment unexpectedly already exists
If a CnsNodeVmAttachment already exists when we were expecting to create
it, inspect the existing one to see if we're the owners of it but it is
otherwise stale. The usual cause of this will be that VC VM was deleted
from underneath, and the replacement VM will end up with a new BiosUUID.

This addresses the immediate issue we have, but still leaves some
lingering issues around the whole attachment workflow, like w.r.t.
volume and PVC name changes, and how we determine if the volume is
actually attached, that we'll have to keep throwing code at until
we're able to revisit this workflow, and likely bring the attachment
and detachment in-house.