Initial v1a2 provider
There is still a lot of work left to do before switching over to v1a2
but this is able to create TKGs with everything else still only
understanding v1a1, so we have a good baseline of version conversion
and functionality.

Unfortunately the way the existing code is and our need to support the
existing v1a1 and new v1a2 APIs in the same codebase makes this a large
and painful change. The goal is to wrap up the remaining items as
quickly as possible so we can move to v1a2 and then really start to
focus on some of our long term tech debt so future version bumps aren't
nearly as painful.