api: Rm CLS fields from VMI print cols
This patch removes the printer columns related to Content Library
from the v1a1 VirtualMachineImage and ClusterVirtualMachineImage
resources. This is aligned with v1a2, and frankly, these fields
provide no real value when printing VMIs and just clutter the
output. A user should not care where the image originates, and
if they do, they can look at the image's status for more info.