🌱 Prevent non-admin users from setting first-boot-done annotation (#237)
**What does this PR do, and why is it needed?**
This patch updates the VM mutation webhook to disallow non-admin users from adding/updating/removing the `first-boot-done` annotation. It is only meant to be modified by the VM-Operator. Also, it fixes the current implementation which could cause all VM updates from non-admin users to fail if the `cloud-init-instance-id` annotations exist.

**Which issue(s) is/are addressed by this PR?**
Fixes #221 

**Are there any special notes for your reviewer**:
This change is a prerequisite for another upcoming change that will rely on the existence of the `first-boot-done` annotation set by the VM-Operator.

**Please add a release note if necessary**:
```release-note
Prevent non-admin users from setting first-boot-done annotation
```