Reverse configSpec.firmware order of precedence to Image then Classes (#238)
This change reverses of the order of precedence to populate the configSpec from the image first
then the class. This is necessary with vsphere UI unable to create VM Classes with an empty firmware,
potentially creating scenarios for VM boot failures when the wrong firmware type is populated.
The image is known to always have the supported firmware type for boot.

This reversal can be changed back once vSphere GUI supports creating VM Classes with an optional firmware
type for a user.