Synthesize v1a1 Prereq condition from v1a2 conditions
There is various weirdness around how and what we set the v1a1 Prereq
condition to, so this isn't quite a complete match but hopefully good
enough.

For now just leave the conditions as-is when going from v1a1 to v1a2. We
might later want to remove the Prereq - or other v1a1 only conditions -
during that conversion.