✨ Persist PVC related info in VM ExtraConfig for backup (#257)
This PR stores a VM's disk UUID to PVC mapping info in ExtraConfig during the backup process. This will enable us to register the volumes of the VM with the same PVC name and access modes during the restore. The change has also been added to the v1a2 package.

A helper function is added in vmprovider_vm_utils.go to retrieve the disk UUID to PVC mapping. This eliminates the need to pass a K8s client in the backup.go file. Also, the BackupVirtualMachine function is updated to have all the necessary params from the new BackupOptions struct.