Remove updates to DeviceGroups
This change removes the update of deviceGroups in the UpdateVM path.
We don't currently support this workflow on Update and DeviceGroups are
already covered in the CreatePath. This code for this currently incomplete
to support deviceGroups on update and will be prioritized
when we support Resizing VMs in VM operator.