Make VM Spec BootDiskCapacity field a pointer
This field is a resource.Quantity which when not set still outputs
as "0". This field is not likely commonly used, and outputting of
a zero size is potentially confusing or misleading. Also, in the
k8s API, resource.Quantity fields are always a pointer.

While here, remove this field defaulting in the version conversion
fuzzer since that is now supported after f65ef245.