Update deprecated pr labeler version
Bump the version of the PR Labeler action, updating to the newer
configuration schema. The size configuration was changed a few versions
ago. I'm looking to remove support so I'm sending some PRs to the very
few repos that still rely on it.

I updated the reference to the action to master. This actually uses the
latest stable version so it should avoid any unintended breakage.

Thanks for using the action!

Signed-off-by: Galo Navarro <anglorvaroa@gmail.com>