Add note that the TCPSocket probe method is deprecated
This probe method was added at the last minute of the first WCP release
as a stopgap to address a limitation with the NSX-T LB at the time.
Since the probe is done via VM Operator pod on one of the Supervisor
CP VMs it requires network connectivity that does not exist in certain
environments, and TKGs - the primary user of this probe method - will
be switching to different method that's introduced in v1a2.