Remove old v1a2 helper function in the VM validation webhook
This was needed at the time due to how the v1a2 code was imported.
While here, update a few import paths to use the "correct" provider,
even though it doesn't really matter because the code is the same.
There's a major package reorganization due once we have some breathing
room.