Adds inline sysprep VM bootstrap support
This patch adds support for inline sysprep which was introduced in the
API in v1alpha2. For the secrets, we are currently decoding them and
adding the values SysprepSecretData field under BootstrapArgs.

This patch also adds tests for the changes to enable inline sysprep.

Instead of waiting till doBootstrap to pull the secret data for inline
sysprep, this patch adds the secret data to the BootstrapArgs earlier
for verification purposes.

It updates the API to mark certain fields as mandatory.

This patch moves the defaulting logic of the secret selector keys to the
CRDs instead of relying on a mutating webhook

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>