Remove Optional from SecretKeySelector
This patch updates references to SecretKeySelector to use an
in-repo version of this type that no longer has the ability
to specify Optional. As was noted earlier by a different
engineer on the project, this field does not make sense for
the way this project uses SecretKeySelector. All the places
where SecretKeySelector is used, the requested data is
required.