Fixes VMI/CVMI condition conversion
When the v1a2 FSS is enabled, all VMI and CVMIs are missing the
conditions when requesting a v1a1 object. This patch introduces a fix
wherein when no false conditions are reported during the down conversion
of a VMI (or CVMI), it sets all the child conditions as True.

Also, the latest transition time for True conditions was missing during
the down conversion. To fix that, we set the transition time from the
Ready condition to each of the child conditions since we do not have the
individual times for each of those conditions anymore.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>