Add VirtualMachine hub-spoke-hub and spoke-hub-spoke tests
Start with a nearly fully populated v1a2 VM and ensure that
the VM resulting from conversion is the same. We can keep adding
these explicit conversion tests to ensure that fields are converted
as we expect instead of always depending on the fuzzer. Add typical
v1a1 TKG VM and ensure spoke-hub-spoke conversion works.

Fix a few conversion issues around the Network and Advanced fields.

Always restore the full VM Status from the annotations. Our controller
should be the only one updating the Status, and an update from the
conversion will cause our controllers to reconcile and reevaluate the
Status.