Check for well-known CloudInit userdata keys
v1a1 did not have a Key field to specify the Secret key to find
the userdata in so we'd try a few well-known keys. So for just
CloudInit check for the existence of those well-known keys as
well when checking for a required key in the CloudInit Secret.

This fixes TKGs post 5e4469e9.