Restore previous LinuxPrep default customization behavior
We have E2E tests that assume this behavior so preserve that in v1a2 as
well. The pointer API change in b54465 changed the behavior to exit
bootstrap early. This is just a quick fix to get pipelines fixed. There
is an existing story to both improve this default behavior - like by
inspecting the image OS, and to keep addressing the prior lack of tests
in this area.