Fix generated CloudConfig type via Docker
This patch fixes a weird issue when generate the Golang type(s) for
a Cloud-Init CloudConfig. When generating them via Docker, the
program quicktype used the schema's filename as part of the generated
types. This caused the file to look different when generated locally
versus when generated via a container.