Fix cloudconfig runcmd / YAML->JSON on parse
This patch updates the inline cloud config's runcmd field to be
of type json.RawMessage instead of a slice of them. This is because
kubebuilder cannot generate the correct CRD for elements in the slice,
and they are expected to be base64-encoded. Instead the entire runcmd
is now a json.RawMessage.

Additionally, this patch fixes an issue in the validation logic for
the inline cloud config. It was based on YAML in, YAML out, but it
should have been JSON in, YAML out.