Do not use os.Getenv or pkg/lib for config
This patch removes the usage of both os.Getenv and pkg/lib from
throughout the codebase as a means to configure VM Operator. This
method was fragile and resulted in a serial test design. The new
package pkg/config is used to define the VM Operator configuration
as well as store/retrieve the config from a Go context.