Fix VirtualMachineImage condition version conversion
When going from an originally v1a1 VMI to v1a2 VMI, the condition
may not have a Reason set since it was optional in v1a1. However,
in v1a2 that is now a required field so ensure that field is set.
We use the Status field since we don't have anything better.

Add a bunch of missing condition conversion tests.