Add GuestDeviceName to network interfaces spec
When using cloud-init this allows the device name in the guest to be set
to something different that what the name is in the k8s VM and used for
the corresponding network interface CR name.

This can be useful if desired device name in the guest does not conform
to what k8s allows since Name field is used for part of the interface CR
name.