🌱 Wait for reconfigure task to complete in v1a2 backup (#352)
This PR updates the v1a2 backup code to use the internal Reconfigure call, which waits for the task to complete. This ensures the required EC keys and their values are stored in the VM for backup.
This PR also removes the unused v1a1 backup code. This is a no-op as both v1a2 and b/r features will be enabled in the same release.