Disable DRS PowerOff for VMs w vGPU/DDPIO
It used to be the case that VI Admins could not manage the power state
of VM Service managed VMs on vSphere via API or the HTML5 client (H5C).
This presented a challenge to maintenance mode. If a VM could not be
automatically evacuated from a host entering maintenance mode, the host
could not successfully enter maintenance mode. In vSphere 7, VMs with
vGPUs were not automatically vMotioned, and in vSphere 8+, VMs with
PCI Passthru devices still are not vMotioned. There are also other CPU
related settings that might pin a VM to a host, preventing DRS from
vMotioning a VM to evacuate it off of a host entering maintenance mode.

Normally a VI Admin would power off a VM that blocks maintenance mode,
but since VM Service managed VMs used to disallow VI Admins from doing
so, a work-around was needed. VM Operator would apply a special flag in
a VM's ExtraConfig to inform DRS to power off this VM if a host is
entering maintenance mode. This flag was only applied for VMs with vGPUs
or PCI Passthru devices because these were the only configurations
allowed by VM Operator that would block evacuation-by-vMotion. And
again, at the time this was implemented, VI Admins could not simply
choose to power off these VM Service managed VMs as the VI Admins would
do for any other VM. Since it is not possible to power on a VM on a host
marked for maintenance mode, attempts to power the VM on by VM Operator
will force DRS to attempt to find another compatible host for the VM.

Starting in vSphere 8, VMs with vGPUs *can* be automatically vMotioned,
but VMs managed by VM Service are still powered off due to the special
flag. Also starting in vSphere 8, VM Service managed VMs *can* be
powered off by VI Admins in H5C. Beginning with vSphere 8.0U2, it is
also possible to realize VM Service managed VMs using the *entire*
virtual hardware and configuration surface of a vSphere VM, increasing
the number of ways in which a VM Service managed VM might block a VM
from being evacuated by vMotion.

Due to these factors, it is no longer desirable for DRS to automatically
power off a VM Service managed VM. Instead, if a VM Service VM is
blocking a host from entering maintenance mode, the VI Admin may treat
it like any other VM and power off the VM using the web client or the
vSphere API.