Guest bootstrap status as condition
This patch introduces the condition "GuestBootstrap" such
that its status, reason, and message are all surfaced from reading
the ExtraConfig key "guestinfo.vmservice.bootstrap.condition". The
key's value is comma-delimited, split at most three times:

    0 - status
    1 - reason  (if present)
    2 - message (if present)

This condition allows users specifying run commands using either
the CloudInit or Sysprep bootstrap providers to leverage the
vmware-rpctool or vmtoolsd commands to update the aforementioned
guestinfo key with a value that is ultimately realized in the
Kubernetes VM's status as a condition.