Include required storage class in StorageReady condition
If we require a storage class - which as of today will always be the
case in WCP, include that as a part of the StorageReady condition check.
There is little point in letting the VM creation flow get past this
point since we will be able to create the VM.

There is some future that will make this optional for certain VMs so
we're not going to enforce this in the webhook at the moment. There is
also future UX improvement here: we don't allow the storage class to
change - this will also be relaxed in the future - so the only way out
of this is to delete the k8s VM resource and recreate it with the
intended storage class. We could use the presence of this condition to
allow it to be changed.

Remove API doc mentioning this field being defaulted when there is a
sensible default, pushing this to v1a3+. The uncertainty about how and
when the storage class will be optional for some VMs make me worried
about trying to determine a default this moment, so keep the existing
v1a1 behavior for v1a2.

Also apply #375 that was inadvertently done to the old v1a1 provider
to the new v1a2 provider.