Created-at annotations
This patch introduces support for the "created-at"
annotations for VM resources. A VM's mutation webhook
ensures the first time a VM is created the following
annotations are patched into the VM:

    * vmoperator.vmware.com/created-at-build-version
    * vmoperator.vmware.com/created-at-schema-version

The above information allows future changes to VM Operator
know helpful details about the version of VM Operator and schema
version that were used to create a VM resource. Even
the absence of the information is informative as it indicates
the VM was created before the annotations were introduced.