Commonize setting of optional VM ready conditions
The VM SetResourcePool, Bootstrap, and Network are optional so only set
their corresponding ready conditions when they're specified.  We were
setting the Bootstrap ready condition when Bootstrap was nil. We we were
not setting the other two when not specified or disabled.  Add some
conditions.Delete() calls to make the not setting a condition explicit.

Add a bunch of expected condition assertions