Set HW Version in CreateConfigSpec()
The logic for determining the ConfigSpec version was spread out in two
different functions and done separately after creating the ConfigSpec.
The version can come from the class ConfigSpec or the class (passthru)
devices, image, VM Spec MinHardwareVersion, or if the VM has PVCs at
create time. Now have one function that is responsible for determining
the version.