Disallow empty sysPrep bootstrap
If neither the raw or cooked Sysprep data is provided the Windows GOSC
Customize() call will fail, and since an empty text for raw Sysprep also
fails just disallow that configuration.

If wanting to do an empty Sysprep ends up being a common request - so
the networking is configured - we can investigate more to see if there
is some noop default we can fallback to.