Remove backing-less VirtualEthernetCards from placement ConfigSpec
The VC placement call PlaceVmsXCluster() cannot cope with
VirtualEthernetCards that do not have a backing. Since a backing-less
card cannot really influence the placement decision remove them from the
ConfigSpec used for placement. The underlying bug appears to only impact
VirtualEthernetCards.

Backingless VirtualEthernetCards can occur in NSX-T since we cannot
resolve the DVPG from the LSX-T LogicalSwitchUUID until we know the CCR.
The backing is assigned after placement when we know the CCR so the VM
is created on the correct DVPG.