Refactor vSphere provider client to pkg/util/vsphere/client
This patch refactors the vSphere provider client into a non-provider
package. The original provider client is now a composite object that
uses the new client. All of the tests from the old package have been
copied to the new package, but left in the original as well to prove
nothing is broken with the drop-in replacement.

Additionally, the client tests now work with the race flag enabled
due to Ginkgo2's ordered ability.

This allows the vSphere client to be easily reused by other locations
outside of the provider.