Add Sysprep bootstrap secret data in createArgs & updateArgs (#382)
This seems like a miss in https://github.com/vmware-tanzu/vm-operator/commit/a8c499d14118a0b9c687d850135fe1594121b7ce when we added the inline Sysprep bootstrap support.
If a Secret key is specified in the inline Sysprep bootstrap, it does get retrieved, but it's not currently being passed to the create/updateArgs. This causes the `CustomizationGuiUnattended` failed because of missing data.