Make the VM network interfaces immutable
In VDS we have difficulty in matching up the network interfaces in our
VM spec with the VM's devices. Unlike NCP/NSX-T, VDS does not generate
the MAC address - we defer to VC to generate it - and an overly
restrictive limitation that on VDS, the EthCard ExternalID cannot be set
on device create (but can later set it via an edit). So for now, try to
constrain things a bit by making our network interfaces immutable so we
can depend on the device order during create.

Remove the general limitation that only network fields could be changed
when the VM was powered off. We cannot atomically determine if the VM is
off in the first place so we were relying on the Spec field. This is a
general issue that we've never firmly specified the behavior around
either Reconfigures or customization on subsequent power off to on
transitions.

Remove Name as the VM's Status.Network.Interfaces as the map key to make
it optional. It will be unset when there is no matching interface from
the VM's Spec.Network.Interfaces list.

Limit network interfaces to 10 since that is the VC limit