🌱 Update VM backup to avoid unnecessary reconfigures (#386)
This PR updates the backup code to check `spec` (by `metadata.generation`), `labels`, and `annotations` field changes of VM CR to determine if the existing backup is up-to-date. Previously, it used `resourceVersion`, which could lead to unnecessary VM reconfigurations whenever the VM status was updated (e.g. VM readiness failed and its "lastTransitionTime" was updated in each no-op reconciliation).

The CloudInit instance-id backup is removed as part of this patch as the corresponding restore workflow doesn't require it anymore.