Only update CloudInit ConfigSpec if updates are required
If the VM already has the expected vAppConfig - either no vAppConfig or
already has the OvfEnvironmentTransport - depending on the type of
CloudInit we're doing, then don't update the ConfigSpec. If something
fails after the reconfigure, we'd otherwise create needless no-op
Reconfigure tasks.

Consistently use HaveKeyWithValue() in the tests. An improvement for
later is to get rid of the magic gzip+base64 magic strings.