Only default to LinuxPrep GOSC on Linux VMs
In the old v1a1 code, we'd always fallback to LinuxPrep regardless if
the guest was Linux or not, and we have tests that assume that behavior.
Improve that here by only doing that on guests that would indicate to be
some known Linux distro.

Restore part of the bootstrap defaulting doc comment that was removed in
4bdee330.

If we need to support other defaulting for like Windows, we can switch
to using the EnvBrowser to get the OS Family