⚠️ Remove v1a1 code (#401)
This removes the bulk of the v1a1 code - webhooks, controllers, and the old provider.

The prior patch and conditions packages have been moved into the v1a1 webconsolerequest controller since it is the only user of those packages now. That controllers needs to be kept around since the v1a2 name is difference and hence version conversion is not possible.

Remove pre Image Registry code from the vcsim test context. The Image Registry FFS was piggybacking on the v1a2 FFS so just always set the v1a2 config value to true.

The new vsphere provider is left under vsphere2 for now since we plan to decompose that big package to make future API version not as invasive.