Fix PodVMOnStretchedSupervisor with WFFC storage class
With a WaitForFirstConsumer storage class, "-wffc" is appended to the
storage class name but the StorageQuotaPolicy still has just the base
name so we cannot depend on those names always being the same.

Instead list all the StorageQuotaPolicys in the namespace to determine
if the storage class is associated with the namespace.

When the PodVMOnStretchedSupervisor FFS is enabled, both the
ResourceQuotas and StorageQuotaPolicys on non-stretched SV so
drop the AZ > 1 check.