Test Labels
This patch introduces labels to Ginkgo tests. For example, the
following labels now exist:

* api        -- related to the APIs
* controller -- a controller test
* create     -- a webhook create test
* delete     -- a webhook delete test
* envtest    -- depends on envtest
* fuzz       -- related to fuzz testing
* mutation   -- a mutation webhook test
* nsxt       -- related to NSX-T
* update     -- a webhook update test
* validation -- a validation webhook test
* v1alpha1   -- related to v1alpha1
* v1alpha2   -- related to v1alpha2
* vcsim      -- depends on vcsim
* webhook    -- a webhook test

The above labels make it trivial to run tests related to specific
categories, or excluding others. For example, the following command
will run all webhook tests that do not depend on envtest:

    ginkgo -r -v --label-filter '!envtest && webhook'

The flags `--enable-integration-tests` and `--enable-unit-tests` have
also been removed, including their associated environment variables,
`ENABLE_INTEGRATION_TESTS` and `ENABLE_UNIT_TESTS`. There is no longer
an equivalent for disabling all "unit" tests. However, there is a direct
correlation between what used to be considered integration tests and the
`envtest` label. Therefore, to run all the tests that used to be
considered integration tests, use the following command:

    ginkgo -r -v --label-filter 'envtest'

To run all the tests that used to be considered *not* integration tests,
use the following command:

    ginkgo -r -v --label-filter '!envtest'