Misc updates to infra controllers and test labels
Add a note around the infra negative tests that even if the objects
are reconciled the tests will still pass. Note that the default
timeout for these tests is 100ms which in my local testing has been
long enough for them to be reconciled when removing the objects from
the Managers Cache.ByObject map.

Remove vcsim from the controller and webhook labels since these just
tests against the fake VM Provider. Change the old WebConsoleRequest
webhook test label to be v1alpha1.