Don't watch PVCs in non-InstanceStorage envs
For envs when InstanceStorage is not enabled (VDS) there is no point in
watching PVCs since that watch is only needed for when we're waiting
for the IS PVCs to get bound.

For envs when InstanceStorage in enabled (NSX-T) InstanceStorage volumes
generally are not used used but we're still watching PVCs. This change
may be superseded by a later change that only starts the Watch() when we
first encounter a VM that has InstanceStorage volumes.