Fix inability to get ConfigMap/Secret objs
This patch addresses a bug created by the move to the new, built-in
cache logic in Controller-Runtime v0.17.x. In adopting this logic
a mistake was made regarding how the logic corresponded to bypassing
the cache for certain object types. This patch fixes the issue by
once again disabling the global cache for ConfigMap and Secret
resources, but this time using specific caches for the infra
controllers.

A new test package, pkg/manager/test, has also been created to
verify that the global cache is in fact bypassed for the disabled
objects.