Reduce VM props fetching in steady state
During VM reconcile, we go to VC to fetch the VM properties twice. Once
to see if anything in the VM's HW config needed to be updated, and again
to populate the VM Status.

In the first fetch do not get "runtime" but instead use the
"summary.runtime" since we were already fetching the summary. These two
runtime's use the same struct.

In the common case - say the VM is already on or off - no changes to the
VM config are done, so the properties fetched first are still entirely
valid. Reuse those properties in that case.

CPU profiling showed that we'd spend a lot of time demarshalling the
response, plus the latency of going to VC, this would slow down our VM
reconcile since we have a set pool of workers.

There is still a lot of room for improvement here to be smarter about
what properties are fetched based on the VM desired state.  I've got
more work planned in this area for later feature work that will improve
upon this.

Rename "vmMO *mo.VirtualMachine" to moVM to keep naming consistent