Fix "make generate" due to testlabels mod
This patch fixes the "make generate" target, which was
failing due to the introduction of the testlabels module.
Apparently this was not caught by the PR testing, so we
need to look into that. The controller-gen binary was not
happy with the ./pkg/... path that included another module,
i.e. ./pkg/constants/testlabels. Instead the command has
been updated to specify a single path, the recursive pattern
for the main module, which results in scanning more paths
than before, but importantly excludes the testlabels module.