Sync VMI status to labels
This patch ensures the VirtualMachineImage and ClusterVMI
resources have their status.capabilities and status.osInfo
fields set on the resources as labels. This behavior was
already documented via the v1alpha2 APIs, but was not
implemented. This change allows clients to more efficiently
discover images based on labels, indexed by the API server,
for information about the images' OS or capabilities.